"Refactoring: get request to resource"
