+++ python_pytest_api/test/pytest_api_rest_get_task_one.py
@@ -3,6 +3,7 @@ import json
 import requests
 import pytest
 from resource_dir import resourse_data
+from resource_dir.requests import get_pair_data
 
 
 TESTED_PAIR = 'USDRUB'
@@ -17,11 +18,11 @@ def expected_code():
     return {'OK': 200, 'NotKnownPair': 1002}
 
 
-def get_pair_data(endpoint):
-    get_request = requests.get(endpoint,
-                               headers={'Content-Type': 'application/json'})
-    requested_data = json.loads(get_request.text)
-    return requested_data
+# def get_pair_data(endpoint):
+#     get_request = requests.get(endpoint,
+#                                headers={'Content-Type': 'application/json'})
+#     requested_data = json.loads(get_request.text)
+#     return requested_data
 
 
 def test_code_exists():
