+++ python_pytest_api/test/pytest_api_rest_get_task_one_fixture.py
@@ -3,6 +3,7 @@ import json
 import requests
 import pytest
 from resource_dir import resourse_data
+from resource_dir.requests import get_pair_data
 
 TESTED_PAIR = 'USDRUB'
 ENV_URL = resourse_data.BASE_URL
@@ -23,11 +24,11 @@ def introduction():
     print('Test is finished')
 
 
-def get_pair_data(endpoint):
-    get_request = requests.get(endpoint,
-                               headers={'Content-Type': 'application/json'})
-    requested_data = json.loads(get_request.text)
-    return requested_data
+# def get_pair_data(endpoint):
+#     get_request = requests.get(endpoint,
+#                                headers={'Content-Type': 'application/json'})
+#     requested_data = json.loads(get_request.text)
+#     return requested_data
 
 
 @pytest.mark.usefixtures('introduction', 'setup')
