+++ python_pytest_api/test/pytest_api_rest_get_task_three.py
@@ -2,6 +2,7 @@
 import json
 import requests
 from resource_dir import resourse_data
+from resource_dir.requests import get_pair_data
 
 
 TESTED_PAIR = 'NOT_EXISTING_PAIR'
@@ -10,11 +11,11 @@ ENDPOINT = ENV_URL + TESTED_PAIR
 ERROR_MESSAGE = 'was not recognised or supported'
 
 
-def get_pair_data(endpoint):
-    get_request = requests.get(endpoint,
-                               headers={'Content-Type': 'application/json'})
-    requested_data = json.loads(get_request.text)
-    return requested_data
+# def get_pair_data(endpoint):
+#     get_request = requests.get(endpoint,
+#                                headers={'Content-Type': 'application/json'})
+#     requested_data = json.loads(get_request.text)
+#     return requested_data
 
 
 def test_message_contains_error_text():
