+++ python_pytest_api/test/pytest_api_rest_get_tasks_parameterized.py
@@ -3,6 +3,7 @@ import json
 import requests
 import pytest
 from resource_dir import resourse_data
+from resource_dir.requests import get_pair_data
 
 
 ENV_URL = resourse_data.BASE_URL
@@ -15,11 +16,11 @@ RESPONSE_CODE_LIST = [
     ]
 
 
-def get_pair_data(endpoint):
-    get_request = requests.get(endpoint,
-                               headers={'Content-Type': 'application/json'})
-    requested_data = json.loads(get_request.text)
-    return requested_data
+# def get_pair_data(endpoint):
+#     get_request = requests.get(endpoint,
+#                                headers={'Content-Type': 'application/json'})
+#     requested_data = json.loads(get_request.text)
+#     return requested_data
 
 
 @pytest.mark.check_response_code
