+++ python_pytest_api/test/resource_dir/requests.py
@@ -1,0 +1,9 @@
+import json
+import requests
+
+
+def get_pair_data(endpoint):
+    get_request = requests.get(endpoint,
+                               headers={'Content-Type': 'application/json'})
+    requested_data = json.loads(get_request.text)
+    return requested_data
